requirements.txt
apache-beam==2.58.0
boto3==1.34.0
pyspark==3.5.0
apache-airflow==2.9.0
mlflow==2.13.0
streamlit==1.35.0
openlineage-python==1.0.0
apache-airflow-providers-openlineage==2.4.0
pandas==2.2.2
pyarrow==16.1.0